Rename camera component
